requirements.txt
streamlit
tensorflow
Pillow
numpy
langchain
bs4
requests
unstructured
